Narrow down functionality of abstract `Annotator`.
