+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/annotation/Annotator.java
@@ -35,12 +35,9 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayDeque;
-import java.util.Queue;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.util.Exceptions.illegalStateWithCauseOf;
-import static io.spine.util.Exceptions.newIllegalStateException;
 import static java.nio.file.Files.exists;
 import static java.nio.file.StandardOpenOption.TRUNCATE_EXISTING;
 
@@ -71,10 +68,6 @@ public abstract class Annotator {
      */
     protected final void annotate(SourceFile relativeSourcePath) {
         rewriteSource(relativeSourcePath, new TypeDeclarationAnnotation());
-    }
-
-    protected final void annotate(ClassName targetClass) {
-        rewriteSource(targetClass.resolveFile(), new NestedTypeDeclarationAnnotation(targetClass));
     }
 
     /**
@@ -151,10 +144,6 @@ public abstract class Annotator {
         }
     }
 
-    protected final Path generatedProtobufDir() {
-        return genProtoDir;
-    }
-
     /**
      * An annotation function, that annotates the type declaration,
      * which is represented by {@link AbstractJavaSource}.
@@ -165,50 +154,6 @@ public abstract class Annotator {
         public void accept(@Nullable AbstractJavaSource<JavaClassSource> input) {
             checkNotNull(input);
             addAnnotation(input);
-        }
-    }
-
-    private final class NestedTypeDeclarationAnnotation implements SourceVisitor<JavaClassSource> {
-
-        private final ClassName targetClass;
-
-        private NestedTypeDeclarationAnnotation(ClassName targetClass) {
-            this.targetClass = targetClass;
-        }
-
-        @Override
-        public void accept(AbstractJavaSource<JavaClassSource> source) {
-            checkNotNull(source);
-            AnnotationTargetSource<?, ?> target = findTarget(source);
-            addAnnotation(target);
-        }
-
-        private AnnotationTargetSource<?, ?> findTarget(AbstractJavaSource<JavaClassSource> root) {
-            String className = targetClass.value();
-            Queue<AbstractJavaSource<JavaClassSource>> classesToCheck = new ArrayDeque<>();
-            classesToCheck.add(root);
-            while (!classesToCheck.isEmpty()) {
-                AbstractJavaSource<JavaClassSource> currentClass = classesToCheck.poll();
-                if (currentClass.getQualifiedName().equals(className)) {
-                    return currentClass;
-                }
-                if (currentClass.hasNestedType(className)) {
-                    return currentClass.getNestedType(className);
-                }
-                currentClass.getNestedTypes()
-                            .stream()
-                            .filter(type -> type instanceof AbstractJavaSource)
-                            .forEach(type -> {
-                                @SuppressWarnings("unchecked")
-                                    // Due to inconvenience of Roaster API.
-                                AbstractJavaSource<JavaClassSource> abstractJavaSource =
-                                        (AbstractJavaSource<JavaClassSource>) type;
-                                classesToCheck.add(abstractJavaSource);
-                            });
-            }
-            throw newIllegalStateException("Class `%s` not found in Java source `%s`.",
-                                           className,
-                                           root.getCanonicalName());
         }
     }
 }
